chore(package): Refactoring all packages architecture.
