+++ app/src/main/java/com/lebaillyapp/uiwavedeformation/ui/components/phaseII/modeA/DeformableBitmapGridTouch.kt
@@ -1,5 +1,8 @@
-package com.lebaillyapp.uiwavedeformation.ui.components.phaseII
+package com.lebaillyapp.uiwavedeformation.ui.components.phaseII.modeA
 
+import android.graphics.Paint
+import android.graphics.Rect
+import android.graphics.RectF
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.gestures.awaitFirstDown
 import androidx.compose.foundation.gestures.forEachGesture
@@ -103,10 +106,10 @@ fun DeformableBitmapGridTouch(
         val cellHeight = drawHeight / gridRows
 
         val androidBitmap = bitmap.asAndroidBitmap()
-        val paint = android.graphics.Paint().apply { isFilterBitmap = true }
+        val paint = Paint().apply { isFilterBitmap = true }
 
-        val srcRect = android.graphics.Rect()
-        val dstRect = android.graphics.RectF()
+        val srcRect = Rect()
+        val dstRect = RectF()
 
         val nativeCanvas = drawContext.canvas.nativeCanvas
