+++ app/src/main/java/com/lebaillyapp/uiwavedeformation/ui/screen/phaseII/modeA/AdaptiveTilteDemoScreen.kt
@@ -1,11 +1,11 @@
-package com.lebaillyapp.uiwavedeformation.ui.screen.phaseII
+package com.lebaillyapp.uiwavedeformation.ui.screen.phaseII.modeA
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.lebaillyapp.uiwavedeformation.ui.components.phaseII.WaveDeformableBitmapGridHybrid
-import com.lebaillyapp.uiwavedeformation.viewmodel.phaseII.WaveTileViewModel
+import com.lebaillyapp.uiwavedeformation.ui.components.phaseII.modeA.WaveDeformableBitmapGridHybrid
+import com.lebaillyapp.uiwavedeformation.viewmodel.phaseII.modeA.WaveTileViewModel
 
 @Composable
 fun AdaptiveTileDemo(bitmap: ImageBitmap,modifier: Modifier = Modifier){
