+++ app/src/main/java/com/lebaillyapp/uiwavedeformation/ui/screen/phaseII/modeA/DemoWaveGridScreen.kt
@@ -1,11 +1,11 @@
-package com.lebaillyapp.uiwavedeformation.ui.screen.phaseII
+package com.lebaillyapp.uiwavedeformation.ui.screen.phaseII.modeA
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ImageBitmap
-import com.lebaillyapp.uiwavedeformation.ui.components.phaseII.WaveDeformableBitmapGrid
-import com.lebaillyapp.uiwavedeformation.viewmodel.phaseII.WaveTileViewModel
+import com.lebaillyapp.uiwavedeformation.ui.components.phaseII.modeA.WaveDeformableBitmapGrid
+import com.lebaillyapp.uiwavedeformation.viewmodel.phaseII.modeA.WaveTileViewModel
 
 @Composable
 fun DemoWaveGridScreen(bitmap: ImageBitmap,
