+++ app/src/main/java/com/lebaillyapp/uiwavedeformation/ui/screen/phaseII/modeB/WaterEffectComposable.kt
@@ -1,54 +1,30 @@
-package com.lebaillyapp.uiwavedeformation.ui.shader
+package com.lebaillyapp.uiwavedeformation.ui.screen.phaseII.modeB
 
 import android.graphics.BitmapFactory
+import android.graphics.Paint
 import android.os.Build
 import android.os.SystemClock
 import androidx.annotation.DrawableRes
 import androidx.annotation.RawRes
 import androidx.annotation.RequiresApi
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.core.rememberInfiniteTransition
 import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.gestures.detectDragGestures
-import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.setValue
 import androidx.compose.runtime.withFrameNanos
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
 import androidx.compose.ui.graphics.nativeCanvas
-import androidx.compose.ui.input.pointer.PointerEventPass
 import androidx.compose.ui.input.pointer.PointerEventType
-import androidx.compose.ui.input.pointer.PointerInputChange
-import androidx.compose.ui.input.pointer.changedToDown
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.IntSize
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.lebaillyapp.uiwavedeformation.model.WaveParams
-import com.lebaillyapp.uiwavedeformation.viewmodel.phaseII.WaveTrailViewModel
-import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.sample
-import kotlinx.coroutines.yield
-import kotlin.coroutines.cancellation.CancellationException
-
-
+import com.lebaillyapp.uiwavedeformation.ui.shader.WaveDeformMultiShaderBrushProvider
 
 
 @RequiresApi(Build.VERSION_CODES.TIRAMISU)
@@ -149,7 +125,7 @@ fun WaterEffectComposable(
 
         drawIntoCanvas { canvas ->
             val shader = brush.createShader(size)
-            val paint = android.graphics.Paint().apply { this.shader = shader }
+            val paint = Paint().apply { this.shader = shader }
             canvas.nativeCanvas.drawRect(0f, 0f, size.width, size.height, paint)
         }
     }
